# Faster determinant for matrices over gf2e (M4RIE)