--- a/src/sage/matrix/matrix_gf2e_dense.pyx
+++ b/src/sage/matrix/matrix_gf2e_dense.pyx
@@ -1669,7 +1669,6 @@ cdef class Matrix_gf2e_dense(matrix_dense.Matrix_dense):
         return det
 
 
-
 def unpickle_matrix_gf2e_dense_v0(Matrix_mod2_dense a, base_ring, nrows, ncols):
     r"""
     EXAMPLES::